Move limter to middleware
